+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 3.1.0 (IN PROGRESS)
 
 * [UICONSET-216](https://folio-org.atlassian.net/browse/UICONSET-216) Issues with displaying data export logs from Member tenant in Consortium manager.
+* [UICONSET-228](https://folio-org.atlassian.net/browse/UICONSET-228) Add newly added 'hideUserLink' prop to 'RoleDetails' component to remove link from users.
 
 ## [3.0.0](https://github.com/folio-org/ui-consortia-settings/tree/v3.0.0) (2025-03-13)
 [Full Changelog](https://github.com/folio-org/ui-consortia-settings/compare/v2.0.2...v3.0.0)